refactor(test): split printer tests by feature into __tests__/
Move the 584-line printer.test.ts into separate files under src/__tests__/
(auth, density, print, connection, retransmission) and extract shared
createTestPrinter / simulateNotification helpers to remove duplicated
mock-setup boilerplate. auth/image tests remain co-located.
